backend/requirements.txt
streamlit
google-genai
python-dotenv
requests
huggingface_hub
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
